mgid bid adapters: refactoring for trimmer code (#12057)
* added consts for values

* added copy method

* moved extractDomainFromHost

* added method triggerNurlWithCpm

* added getUserSyncs

* less formatting mgidBid

* less formatting undertoneBid

* less formatting utils

* renamed adm to nativeAdm

* back changes for other adaptes

* back changes

* moved getUserSyncs to lib pkg

* less changes

* fixed errors

* less changes

* fixed errors

* fixed tests

---------
